refacto-webapp: refacto models
