+++ webapp/src/components/TableWrapper/index.tsx
@@ -10,15 +10,12 @@ import {
   Select,
   MenuItem,
   CardHeader,
-  Button
 } from '@mui/material';
-import AddCircleOutlineIcon from '@mui/icons-material/AddCircleOutline';
 import BulkActions from './BulkActions';
-import { StatusEnum, StatusOptions } from 'src/models/status';
-import { addItem, deleteItem, getItems, updateItem } from 'src/services/common.service';
+import { StatusEnum, StatusOptions } from 'src/models';
+import { deleteItem, getItems, updateItem } from 'src/services/common.service';
 import { useModal } from 'src/hooks/useModal';
 import useAuth from 'src/hooks/useAuth';
-import { TableEnum } from 'src/models';
 
 interface TableWrapperProps {
   className?: string;
@@ -29,13 +26,6 @@ interface TableWrapperProps {
   children: ReactNode,
 }
 
-export type TableItem = any;
-
-export interface TableMethods {
-  handleDeleteItem: (id: string) => void,
-  handleUpdateItem: (id: string, data: object) => void,
-  handleGetItems: () => void
-}
 
 export interface TableProps<T> {
   items: T[], 
@@ -49,7 +39,6 @@ export interface TableProps<T> {
   addModalOpen: boolean,
   handleCloseAddModal: () => void,
   handleGetItems: () => void,
-  table?: TableEnum,
   statusOptions: StatusOptions
 }
