+++ webapp/src/hooks/useModal.tsx
@@ -1,5 +1,5 @@
 import { useState } from "react"
-import { TableItem } from "src/components/TableWrapper";
+import { TableItem } from "src/models";
 
 type UseModalResponse = [
     modalOpen: boolean,
