+++ webapp/src/models/admin.ts
@@ -1,14 +1,6 @@
 import { Role } from "./role";
-import { Status, StatusEnum } from "./status";
+import { AdminStatus } from "./status";
 
-export enum AdminStatusEnum {
-    actived = StatusEnum.actived,
-    disabled = StatusEnum.disabled,
-    deleted = StatusEnum.deleted,
-    freezed = StatusEnum.freezed,
-}
-
-export type AdminStatus = Status<keyof typeof AdminStatusEnum>;
 
 export interface Admin {
     id: string,
