--- webapp/src/models/application.ts
@@ -1,6 +1,0 @@
-export interface Application {
-    id: string,
-    name: string,
-    createdAt: string,
-    updatedAt: string
-}
