--- webapp/src/models/attachment_chat.ts
@@ -1,8 +1,0 @@
-export interface AttachmentChat {
-    id: string,
-    original_name: string,
-    type: string,
-    attachment: string,
-    createdAt: string,
-    updatedAt: string
-}
