--- webapp/src/models/attachment_job.ts
@@ -1,8 +1,0 @@
-export interface AttachmentJob {
-    id: string,
-    original_name: string,
-    type: string,
-    image: string,
-    createdAt: string,
-    updatedAt: string
-}
