+++ webapp/src/models/commons.ts
@@ -1,26 +1,9 @@
-import { DiscountType } from "./discount_type";
-import { JobDifficulty } from "./job_difficulty";
-import { LitigationObject } from "./litigation_object";
+import { DiscountType } from "./discount";
+import { JobDifficulty, Tag } from "./job";
+import { LitigationObject } from "./litigation";
 import { Pricing } from "./pricing";
 import { Role } from "./role";
-import { Status, StatusEnum } from "./status";
-import { Tag } from "./tag";
-
-export enum CommonsUriEnum {
-    status = "/status",
-    roles = "/roles",
-    tags = "/tags",
-    difficulties = "/difficulties",
-    litigationObjects = "/litigation-objects",
-    discountTypes = "/types-discounts",
-    appearances = "/appearances",
-    pricings = "/pricings"
-}
-
-export enum CommonStatusEnum {
-    actived = StatusEnum.actived,
-    deleted = StatusEnum.deleted
-}
+import { Status } from "./status";
 
 export type Commons = 
     Status<any>      |
