--- webapp/src/models/discount_type.ts
@@ -1,7 +1,0 @@
-export interface DiscountType {
-    id: string,
-    name: string, 
-    type: string,
-    createdAt: string,
-    updatedAt: string
-}
