+++ webapp/src/models/index.ts
@@ -1,26 +1,19 @@
-export type { Job } from './job';
-export type { JobDifficulty } from './job_difficulty';
+export type { Job, JobDifficulty, AttachmentJob, Tag } from './job';
 export type { Organization } from './organization';
-export type { Litigation } from './litigation';
-export type { LitigationObject } from './litigation_object';
+export type { Litigation, LitigationObject  } from './litigation';
 export type { Role } from './role';
-export type { Tag } from './tag';
-export type { User } from './user';
+export type { User, Membership } from './user';
 export type { Admin } from './admin';
 export type { Status, StatusOptions } from './status';
-export type { Discount } from './discount';
-export type { DiscountType } from './discount_type';
-export type { DiscountTransaction } from './discount_transaction';
+export type { Discount, DiscountType, DiscountTransaction } from './discount';
 export type { Commons } from './commons';
-export type { Chat } from './chat';
-export type { AttachmentChat } from './attachment_chat';
-export type { AttachmentJob } from './attachment_job';
+export type { Chat, AttachmentChat } from './chat';
 export type { Partner } from './partner';
 export type { Pricing } from './pricing';
 export type { Subscription } from './subscription';
-export type { Membership } from './membership';
+export type { TableItem, TableMethods } from './table'
+export type { Token, LoginData, SignUpData, ErrorResponse } from './services'
 
-export { CommonsUriEnum } from './commons';
+export { BaseUrlEnum, CommonsBaseUrlEnum } from './services';
 export { StatusEnum } from './status';
 export { LangEnum } from './lang';
-export { TableEnum } from './table';
