+++ webapp/src/models/litigation.ts
@@ -1,11 +1,7 @@
-// export type LitigationStatus = 'completed' | 'pending' | 'coming';
-
 import { Job } from "./job";
-import { LitigationObject } from "./litigation_object";
-import { Status, StatusEnum } from "./status";
+import { LitigationObjectStatus, LitigationStatus } from "./status";
 import { User } from "./user";
 
-export type LitigationStatus = Status<StatusEnum.opened | StatusEnum.closed>;
 
 export interface Litigation {
     id: string,
@@ -24,4 +20,14 @@ export interface Litigation {
     createdAt: string,
     updatedAt: string
 }
+
+export interface LitigationObject {
+    id: string,
+    label: string,
+    description: string,
+    status: LitigationObjectStatus,
+    createdAt: string,
+    updatedAt: string
+}
+
