+++ webapp/src/models/organization.ts
@@ -1,15 +1,7 @@
 import { Job } from "./job";
-import { Status, StatusEnum } from "./status";
+import { OrganizationStatus } from "./status";
 import { User } from "./user";
 
-export enum OrganizationStatusEnum {
-    actived = StatusEnum.actived,
-    disabled = StatusEnum.disabled,
-    deleted = StatusEnum.deleted,
-    freezed = StatusEnum.freezed
-}
-
-export type OrganizationStatus = Status<keyof typeof OrganizationStatusEnum>;
 
 export interface Organization {
     id: string,
