+++ webapp/src/models/table.ts
@@ -1,19 +1,7 @@
-export enum TableEnum {
-    admins = "Admins",
-    users = "Users",
-    organizations = "Organizations",
-    litigationObjects = "LitigationObjects",
-    partners = "Partners",
-    jobs = "Jobs",
-    tags = "Tags",
-    pricings = "Pricings",
-    subscriptions = "Subscriptions",
-    litigations = "Litigations",
-    logs = "Logs",
-    discounts = "Discounts",
-    discountTypes = "DiscountTypes",
-    roles = "Roles",
-    status = "Status",
-    difficulties = "Difficulties"
-}
+export type TableItem = any;
 
+export interface TableMethods {
+  handleDeleteItem: (id: string) => void,
+  handleUpdateItem: (id: string, data: object) => void,
+  handleGetItems: () => void
+}
