--- webapp/src/models/tag.ts
@@ -1,13 +1,0 @@
-import { Status, StatusEnum } from "./status";
-
-export type SubscriptionStatus = Status<StatusEnum.actived | StatusEnum.deleted>;
-
-
-export interface Tag {
-    id: string,
-    label: string,
-    type: 'mission' | 'category'
-    status: SubscriptionStatus,
-    createdAt: string,
-    updatedAt: string,
-}
