+++ webapp/src/models/user.ts
@@ -1,18 +1,17 @@
-import { Membership } from "./membership";
 import { Organization } from "./organization";
 import { Partner } from "./partner";
 import { Role } from "./role";
-import { Status, StatusEnum } from "./status";
+import { MembershipStatus, UserStatus } from "./status";
 
 
-export enum UserStatusEnum {
-    actived = StatusEnum.actived,
-    disabled = StatusEnum.disabled,
-    deleted = StatusEnum.deleted,
-    freezed = StatusEnum.freezed
+export interface Membership {
+    id: string,
+    organization: Organization,
+    role: Role,
+    status: MembershipStatus,
+    createdAt: string,
+    updatedAt: string
 }
-
-export type UserStatus = Status<keyof typeof UserStatusEnum>;
 
 export interface User {
     id: string,
